# perf: improve performance, reduce allocations, and avoid promises