@pi0 2024-07-08
Thanks for your efforts on this dear @negezor it looks all good improvements ❤️ Sure i will carefully review once i had time (please read [this](https://github.com/pi0/tired-maintainer) for some understanding)